% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_env.R
\name{add.set_env}
\alias{add.set_env}
\title{Title Add Object to Set}
\usage{
add.set_env(set, object)
}
\arguments{
\item{set}{The set_env to which to add the object.}

\item{object}{The object to add to the set.}
}
\value{
The set_env with the object added.
}
\description{
Title Add Object to Set
}
\examples{
set <- set_env()
obj <- "apple"
set <- add.set_env(set, obj)
}
