% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_env.R
\name{add_many.set_env}
\alias{add_many.set_env}
\title{Title Add Multiple Objects to Set}
\usage{
add_many.set_env(set, items)
}
\arguments{
\item{set}{The set_env to which to add the objects.}

\item{items}{A vector or list of objects to add to the set.}
}
\value{
The set_env with the objects added.
}
\description{
Title Add Multiple Objects to Set
}
\examples{
set <- set_env()
items <- c("apple", "banana", "cherry")
set <- add_many.set_env(set, items) # Should add all items to the set
}
