% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_env.R
\name{as.set_env}
\alias{as.set_env}
\title{Title Convert Object to set_env}
\usage{
as.set_env(object)
}
\arguments{
\item{object}{An object to convert to a set_env.}
}
\value{
A set_env object containing the elements of the input object.
}
\description{
Title Convert Object to set_env
}
\examples{
obj <- c("apple", "banana")
set <- as.set_env(obj) # Should return a set_env with "apple" and "banana"
}
