% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{cls}
\alias{cls}
\title{Inspect objects}
\usage{
cls(x)
}
\arguments{
\item{x}{a collection object}
}
\description{
\code{cls} is a replacement for the \code{class} function
which also works for the collection objects. It falls back to the ordinary \code{class} function
for other objects.
}
\examples{
d <- dict()
cls(d)
}
