% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\docType{package}
\name{collections-package}
\alias{collections}
\alias{collections-package}
\title{collections: High Performance Container Datatypes for R}
\description{
More about what it does (maybe more than one line)
Use four spaces when indenting paragraphs within the Description.
}
\details{
High Performance Container Datatypes for R
}
