% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_env.R
\name{contains.set_env}
\alias{contains.set_env}
\title{Title Check if Object is in Set}
\usage{
contains.set_env(set, object)
}
\arguments{
\item{set}{The set_env to check.}

\item{object}{The object to check for.}
}
\value{
TRUE if the object is in the set, FALSE otherwise.
}
\description{
Title Check if Object is in Set
}
\examples{
set <- set_env()
obj <- "apple"
set <- add.set_env(set, obj)
contains.set_env(set, obj) # Should return TRUE
}
