% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{deprecated}
\alias{deprecated}
\alias{Deque}
\alias{Dict}
\alias{OrderedDict}
\alias{PriorityQueue}
\alias{Queue}
\alias{Stack}
\title{Deprecated Functions}
\usage{
Deque(...)

Dict(...)

OrderedDict(...)

PriorityQueue(...)

Queue(...)

Stack(...)
}
\arguments{
\item{...}{anything}
}
\description{
Deprecated Functions
}
