% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deque.R
\name{Deque}
\alias{Deque}
\title{Double Ended Queue}
\usage{
Deque(...)
}
\description{
The \code{Deque} function creates a double ended queue.
}
\section{Usage}{

\preformatted{
Deque(items = NULL)
.$push(item)
.$pushleft(item)
.$pop()
.$popleft()
.$peek()
.$peekleft()
.$extend(q)
.$extendleft(q)
.$remove(item)
.$clear()
.$size()
.$as_list()
}
}

\section{Arguments}{

\itemize{
\item \code{items}: initialization list
\item \code{item}: any R object
\item \code{q}: a Deque object
}
}

\examples{
q <- Deque()
q$push("foo")
q$push("bar")
q$pushleft("baz")
q$pop()  # bar
q$popleft()  # baz

q <- Deque(list("foo", "bar"))
q$push("baz")$pushleft("bla")
}
\seealso{
\link{DequeL}
}
