% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_env.R
\name{difference.set_env}
\alias{difference.set_env}
\title{Title Difference of Two Sets}
\usage{
difference.set_env(set1, set2)
}
\arguments{
\item{set1}{First set to compare.}

\item{set2}{Second set to compare.}
}
\value{
A new set_env object containing the difference of the two sets.
}
\description{
Title Difference of Two Sets
}
\examples{
set1 <- set_env()
set1 <- add.set_env(set1, "apple")
set1 <- add.set_env(set1, "banana")
set2 <- set_env()
set2 <- add.set_env(set2, "banana")
differenceSet <- difference.set_env(set1, set2)
}
