% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_env.R
\name{get_keys.set_env}
\alias{get_keys.set_env}
\title{Title Get All Keys in Set}
\usage{
get_keys.set_env(set)
}
\arguments{
\item{set}{The set_env to retrieve keys from.}
}
\value{
A character vector of keys in the set.
}
\description{
Title Get All Keys in Set
}
\examples{
set <- set_env()
set <- add_many.set_env(set, c("apple", "banana"))
keys <- get_keys.set_env(set) # Should return c(digest::digest("apple"), digest::digest("banana"))
}
