% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_env.R
\name{get_objects.set_env}
\alias{get_objects.set_env}
\title{Title Get Objects in Set}
\usage{
get_objects.set_env(set)
}
\arguments{
\item{set}{The set_env to retrieve objects from.}
}
\value{
A list of objects in the set.
}
\description{
Title Get Objects in Set
}
\examples{
set <- set_env()
set <- add_many.set_env(set, c("apple", "banana"))
objects <- get_objects.set_env(set) # Should return list("apple", "banana")
}
