% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_env.R
\name{intersection.set_env}
\alias{intersection.set_env}
\title{Title Intersection of Two Sets}
\usage{
intersection.set_env(set1, set2)
}
\arguments{
\item{set1}{First set to compare.}

\item{set2}{Second set to compare.}
}
\value{
A new set_env object containing the intersection of the two sets.
}
\description{
Title Intersection of Two Sets
}
\examples{
set1 <- set_env()
set1 <- add.set_env(set1, "apple")
set1 <- add.set_env(set1, "banana")
set2 <- set_env()
set2 <- add.set_env(set2, "banana")
intersectionSet <- intersection.set_env(set1, set2)
}
