% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priority_queue.R
\name{PriorityQueue}
\alias{PriorityQueue}
\title{Priority Queue}
\usage{
PriorityQueue(...)
}
\description{
The \code{PriorityQueue} function creates a priority queue (a.k.a heap).
}
\section{Usage}{

\preformatted{
PriorityQueue(items = NULL, priorities = NULL)
.$push(item, priority = 0)
.$pop()
.$clear()
.$size()
.$as_list()
}
}

\section{Argument}{

\itemize{
\item \code{items}: initialization list
\item \code{priorities}: a vector of priorities of the same length of \code{items}
\item \code{item}: any R object
\item \code{priority}: a real number, item with larger priority pops first
}
}

\examples{
q <- PriorityQueue()
q$push("not_urgent")
q$push("urgent", priority = 2)
q$push("not_as_urgent", priority = 1)
q$pop()  # urgent
q$pop()  # not_as_urgent
q$pop()  # not_urgent

q <- PriorityQueue(list("not_urgent", "urgent"), c(0, 2))
q$push("not_as_urgent", 1)$push("not_urgent2")
}
