% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{Queue}
\alias{Queue}
\title{Queue}
\usage{
Queue(...)
}
\description{
The \code{Queue} function creates a queue.
}
\section{Usage}{

\preformatted{
Queue(items = NULL)
.$push(item)
.$pop()
.$peek()
.$clear()
.$size()
.$as_list()
}
}

\section{Argument}{

\itemize{
\item \code{items}: initialization list
\item \code{item}: any R object
}
}

\examples{
q <- Queue()
q$push("first")
q$push("second")
q$pop()  # first
q$pop()  # second

q <- Queue(list("foo", "bar"))
q$push("baz")$push("bla")
}
\seealso{
\link{QueueL}
}
