% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\docType{data}
\name{QueueL}
\alias{QueueL}
\title{Queue (list based)}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
QueueL
}
\arguments{
\item{item}{any R object}
}
\description{
The \code{QueueL} class creates a queue with list backend.
It is recommended for short queue.
\preformatted{
QueueL$new()
QueueL$push(item)
QueueL$pop()
QueueL$size()
QueueL$as_list()
}
}
\examples{
q <- QueueL$new()
q$push("first")
q$push("second")
q$pop()  # first
q$pop()  # second
}
\seealso{
\link{Queue}
}
\keyword{datasets}
