% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_env.R
\name{remove.set_env}
\alias{remove.set_env}
\title{Title Remove Object from Set}
\usage{
remove.set_env(set, object)
}
\arguments{
\item{set}{The set_env from which to remove the object.}

\item{object}{The object to remove from the set.}
}
\value{
The set_env with the object removed.
}
\description{
Title Remove Object from Set
}
\examples{
set <- set_env()
obj <- "apple"
set <- add.set_env(set, obj)
set <- remove.set_env(set, obj)
}
