% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\docType{data}
\name{Stack}
\alias{Stack}
\title{Stack}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Stack
}
\arguments{
\item{item}{any R object}
}
\description{
The \code{Stack} class creates a stack with pairlist backend.
It is recommended for long stack
\preformatted{
Stack$new()
Stack$push(item)
Stack$pop()
Stack$size()
Stack$as_list()
}
}
\examples{
s <- Stack$new()
s$push("first")
s$push("second")
s$pop()  # second
s$pop()  # first
}
\seealso{
\link{StackL}
}
\keyword{datasets}
