% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_env.R
\name{union.set_env}
\alias{union.set_env}
\title{Title Union of Two Sets}
\usage{
union.set_env(set1, set2)
}
\arguments{
\item{set1}{First set to combine.}

\item{set2}{Second set to combine.}
}
\value{
A new set_env object containing the union of the two sets.
}
\description{
Title Union of Two Sets
}
\examples{
set1 <- set_env()
set1 <- add.set_env(set1, "apple")
set2 <- set_env()
set2 <- add.set_env(set2, "banana")
unionSet <- union.set_env(set1, set2)
}
