% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\docType{package}
\name{collections-package}
\alias{collections}
\alias{collections-package}
\title{collections: High Performance Container Datatypes for R}
\description{
Provides high performance container datatypes such as Queue, Stack, Deque, Dict and OrderedDict. Benchmarks <https://randy3k.github.io/collections/articles/benchmark.html> have shown that these containers are asymptotically more efficient than those offered by other packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://randy3k.github.io/arrangements}
}

}
\author{
\strong{Maintainer}: Randy Lai \email{randy.cs.lai@gmail.com}

}
