% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deque.R
\docType{data}
\name{Deque}
\alias{Deque}
\title{Double Ended Queue}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Deque
}
\arguments{
\item{item}{any R object}

\item{q}{a Deque object}
}
\description{
The \code{Deque} class creates a double ended queue with pairlist backend.
It is recommended for long queue.
\preformatted{
Deque$new()
Deque$push(item)
Deque$pushleft(item)
Deque$pop()
Deque$popleft()
Deque$extend(q)
Deque$extendleft(q)
Deque$remove(item)
Deque$size()
Deque$as_list()
}
}
\examples{
q <- Deque$new()
q$push("foo")
q$push("bar")
q$pushleft("baz")
q$pop()  # bar
q$popleft()  # baz
}
\seealso{
\link{DequeL}
}
\keyword{datasets}
