% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deque.R
\docType{data}
\name{Deque}
\alias{Deque}
\title{Double Ended Queue}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Deque
}
\description{
The \code{Deque} class creates a double ended queue with pairlist backend.
It is recommended for long queue.
}
\section{Usage}{

\preformatted{
Deque$new()
Deque$push(item)
Deque$pushleft(item)
Deque$pop()
Deque$popleft()
Deque$extend(q)
Deque$extendleft(q)
Deque$remove(item)
Deque$size()
Deque$as_list()
}
}

\section{Arguments}{

\itemize{
\item \code{item}: any R object
\item \code{q}: a Deque object
}
}

\examples{
q <- Deque$new()
q$push("foo")
q$push("bar")
q$pushleft("baz")
q$pop()  # bar
q$popleft()  # baz
}
\seealso{
\link{DequeL}
}
\keyword{datasets}
