% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\docType{data}
\name{Queue}
\alias{Queue}
\title{Queue}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Queue
}
\arguments{
\item{item}{any R object}
}
\description{
The \code{Queue} class creates a queue with pairlist backend.
It is recommended for long queue.
\preformatted{
Queue$new()
Queue$push(item)
Queue$pop()
Queue$size()
Queue$as_list()
}
}
\examples{
q <- Queue$new()
q$push("first")
q$push("second")
q$pop()  # first
q$pop()  # second
}
\seealso{
\link{QueueL}
}
\keyword{datasets}
