% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{queue}
\alias{queue}
\title{Queue}
\usage{
queue(items = NULL)
}
\arguments{
\item{items}{a list of items}
}
\description{
\code{queue} creates a queue.
}
\details{
Following methods are exposed:
\preformatted{
.$push(item)
.$pop()
.$peek()
.$clear()
.$size()
.$as_list()
.$print()
}
\itemize{
\item \code{item}: any R object
}
}
\examples{
q <- queue()
q$push("first")
q$push("second")
q$pop()  # first
q$pop()  # second

q <- queue(list("foo", "bar"))
q$push("baz")$push("bla")
}
\seealso{
\link{stack} and \link{deque}
}
