% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\docType{data}
\name{Queue}
\alias{Queue}
\title{Queue}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Queue
}
\description{
The \code{Queue} class creates a queue with pairlist backend.
}
\section{Usage}{

\preformatted{
Queue$new(items = NULL)
Queue$push(item)
Queue$pop()
Queue$peek()
Queue$clear()
Queue$size()
Queue$as_list()
}
}

\section{Argument}{

\itemize{
\item \code{items}: initialization list
\item \code{item}: any R object
}
}

\examples{
q <- Queue$new()
q$push("first")
q$push("second")
q$pop()  # first
q$pop()  # second

q <- Queue$new(list("foo", "bar"))
q$push("baz")$push("bla")
}
\seealso{
\link{QueueL}
}
\keyword{datasets}
