% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{Queue}
\alias{Queue}
\title{Queue}
\usage{
Queue(items = NULL)
}
\arguments{
\item{items}{a list of items}
}
\description{
The \code{Queue} function creates a queue.
}
\details{
Following methods are exposed:
\preformatted{
.$push(item)
.$pop()
.$peek()
.$clear()
.$size()
.$as_list()
.$print()
}
\itemize{
\item \code{item}: any R object
}
}
\examples{
q <- Queue()
q$push("first")
q$push("second")
q$pop()  # first
q$pop()  # second

q <- Queue(list("foo", "bar"))
q$push("baz")$push("bla")
}
\seealso{
\link{QueueL}
}
