% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{QueueL}
\alias{QueueL}
\title{QueueL (R implementation)}
\usage{
QueueL(...)
}
\description{
The \code{QueueL} function creates a queue.
Pure R implementation for benchmarking.
}
\section{Usage}{

\preformatted{
QueueL(items = NULL)
.$push(item)
.$pop()
.$peek()
.$clear()
.$size()
.$as_list()
}
}

\section{Argument}{

\itemize{
\item \code{items}: initialization list
\item \code{item}: any R object
}
}

\examples{
q <- QueueL()
q$push("first")
q$push("second")
q$pop()  # first
q$pop()  # second

q <- QueueL(list("foo", "bar"))
q$push("baz")$push("bla")
}
\seealso{
\link{Queue}
}
