% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\docType{data}
\name{QueueL}
\alias{QueueL}
\title{Queue (list based)}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
QueueL
}
\description{
The \code{QueueL} class creates a queue with list backend.
Pure R implementation, mainly for benchmark.
}
\section{Usage}{

\preformatted{
QueueL$new(...)
QueueL$push(item)
QueueL$pop()
QueueL$peek()
QueueL$clear()
QueueL$size()
QueueL$as_list()
}
}

\section{Argument}{

\itemize{
\item \code{...}: initialization list
\item \code{item}: any R object
}
}

\examples{
q <- QueueL$new()
q$push("first")
q$push("second")
q$pop()  # first
q$pop()  # second

q <- QueueL$new(list("foo", "bar"))
q$push("baz")$push("bla")
}
\seealso{
\link{Queue}
}
\keyword{datasets}
