% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{Stack}
\alias{Stack}
\title{Stack}
\usage{
Stack(...)
}
\description{
The \code{Stack} function creates a stack.
}
\section{Usage}{

\preformatted{
Stack(items = NULL)
.$push(item)
.$pop()
.$peek()
.$clear()
.$size()
.$as_list()
}
}

\section{Argument}{

\itemize{
\item \code{items}: initialization list
\item \code{item}: any R object
}
}

\examples{
s <- Stack()
s$push("first")
s$push("second")
s$pop()  # second
s$pop()  # first

s <- Stack(list("foo", "bar"))
s$push("baz")$push("bla")
}
\seealso{
\link{StackL}
}
