% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{Stack}
\alias{Stack}
\title{Stack}
\usage{
Stack(items = NULL)
}
\arguments{
\item{items}{a list of items}
}
\description{
The \code{Stack} function creates a stack.
}
\details{
Following methods are exposed:
\preformatted{
.$push(item)
.$pop()
.$peek()
.$clear()
.$size()
.$as_list()
.$print()
}
\itemize{
\item \code{item}: any R object
}
}
\examples{
s <- Stack()
s$push("first")
s$push("second")
s$pop()  # second
s$pop()  # first

s <- Stack(list("foo", "bar"))
s$push("baz")$push("bla")
}
\seealso{
\link{StackL}
}
