% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{stack}
\alias{stack}
\title{Stack}
\usage{
stack(items = NULL)
}
\arguments{
\item{items}{a list of items}
}
\description{
\code{stack} creates a stack.
}
\details{
Following methods are exposed:
\preformatted{
.$push(item)
.$pop()
.$peek()
.$clear()
.$size()
.$as_list()
.$print()
}
\itemize{
\item \code{item}: any R object
}
}
\examples{
s <- stack()
s$push("first")
s$push("second")
s$pop()  # second
s$pop()  # first

s <- stack(list("foo", "bar"))
s$push("baz")$push("bla")
}
\seealso{
\link{queue} and \link{deque}
}
