% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{StackL}
\alias{StackL}
\title{StackL (R implementation)}
\usage{
StackL(...)
}
\description{
The \code{StackL} function creates a stack.
Pure R implementation for benchmarking.
}
\section{Usage}{

\preformatted{
StackL(items = NULL)
.$push(item)
.$pop()
.$peek()
.$clear()
.$size()
.$as_list()
}
}

\section{Argument}{

\itemize{
\item \code{items}: initialization list
\item \code{item}: any R object
}
}

\examples{
s <- StackL()
s$push("first")
s$push("second")
s$pop()  # second
s$pop()  # first

s <- stack(list("foo", "bar"))
s$push("baz")$push("bla")
}
\seealso{
\link{Stack}
}
