% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\docType{data}
\name{StackL}
\alias{StackL}
\title{Stack (list based)}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
StackL
}
\description{
The \code{StackL} class returns a stack with list backend.
It is recommended for short stacks.
}
\section{Usage}{

\preformatted{
StackL$new()
StackL$push(item)
StackL$pop()
StackL$size()
StackL$as_list()
}
}

\section{Argument}{

\itemize{
\item \code{item}: any R object
}
}

\examples{
s <- StackL$new()
s$push("first")
s$push("second")
s$pop()  # second
s$pop()  # first
}
\seealso{
\link{StackL}
}
\keyword{datasets}
