% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{StackL}
\alias{StackL}
\title{StackL (R implementation)}
\usage{
StackL(items = NULL)
}
\arguments{
\item{items}{a list of items}
}
\description{
The \code{StackL} function creates a stack.
Pure R implementation for benchmarking.
}
\details{
Following methods are exposed:
\preformatted{
.$push(item)
.$pop()
.$peek()
.$clear()
.$size()
.$as_list()
.$print()
}
\itemize{
\item \code{item}: any R object
}
}
\examples{
s <- StackL()
s$push("first")
s$push("second")
s$pop()  # second
s$pop()  # first

s <- StackL(list("foo", "bar"))
s$push("baz")$push("bla")
}
\seealso{
\link{Stack}
}
